+++ src/main/java/tech/anonymoushacker1279/iwcompatbridge/plugin/curios/AccessoryBridge.java
@@ -1,19 +1,22 @@
 package tech.anonymoushacker1279.iwcompatbridge.plugin.curios;
 
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.ai.attributes.Attribute;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import tech.anonymoushacker1279.immersiveweapons.event.game_effects.AccessoryManager;
-import tech.anonymoushacker1279.immersiveweapons.item.AccessoryItem;
-import tech.anonymoushacker1279.immersiveweapons.item.AccessoryItem.EffectType;
+import tech.anonymoushacker1279.immersiveweapons.item.accessory.Accessory;
+import tech.anonymoushacker1279.immersiveweapons.item.accessory.AccessoryEffectType;
+import tech.anonymoushacker1279.immersiveweapons.item.accessory.scaling.AccessoryEffectScalingType;
+import tech.anonymoushacker1279.immersiveweapons.item.accessory.scaling.AttributeOperation;
+import tech.anonymoushacker1279.immersiveweapons.item.accessory.scaling.DynamicAttributeOperationInstance;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.SlotResult;
 import top.theillusivec4.curios.api.type.capability.ICuriosItemHandler;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
-import static tech.anonymoushacker1279.immersiveweapons.event.game_effects.AccessoryManager.handleEffectScaling;
 
 /*
  * A class to handle Curios integration. Classes here are invoked by IW in the relevant areas, when IWCB is loaded
@@ -23,24 +26,30 @@ import static tech.anonymoushacker1279.immersiveweapons.event.game_effects.Acces
 public class AccessoryBridge {
 
 	/**
-	 * A modified version of {@link AccessoryManager#collectEffects(EffectType, Player)} designed to work with Curios.
-	 * Collect the value of the given effect from all {@link AccessoryItem}s in the player's inventory.
+	 * A modified version of {@link AccessoryManager#collectEffects(AccessoryEffectType, Player)} designed to work with
+	 * Curios. Collect the value of the given effect from all accessories in the player's inventory.
 	 *
-	 * @param type   the <code>EffectType</code> to collect
+	 * @param type   the <code>AccessoryEffectType</code> to collect
 	 * @param player the <code>Player</code> to collect from
 	 * @return the value of the effect
 	 */
-	public static double collectEffects(EffectType type, Player player) {
+	public static double collectEffects(AccessoryEffectType type, Player player) {
 		double value = 0;
 
 		Optional<ICuriosItemHandler> optional = CuriosApi.getCuriosInventory(player);
 		if (optional.isPresent()) {
 			ICuriosItemHandler itemHandler = optional.get();
-			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItem() instanceof AccessoryItem);
+			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItemHolder().getData(Accessory.ACCESSORY) != null);
 
 			for (SlotResult slotResult : curios) {
-				AccessoryItem accessoryItem = (AccessoryItem) slotResult.stack().getItem();
-				value += handleEffectScaling(accessoryItem, type, player);
+				Accessory accessory = slotResult.stack().getItemHolder().getData(Accessory.ACCESSORY);
+				AccessoryEffectScalingType scalingType;
+				if (accessory != null) {
+					scalingType = accessory.effectScalingTypes().get(type.name());
+					if (scalingType != null) {
+						value += scalingType.getEffectValue(accessory, type, player);
+					}
+				}
 			}
 		}
 
@@ -49,59 +58,59 @@ public class AccessoryBridge {
 
 	/**
 	 * A modified version of {@link AccessoryManager#collectStandardAttributes(Player)} designed to work with Curios.
-	 * Collect the accessory attribute modifiers from all {@link AccessoryItem}s in the player's inventory.
+	 * Collect the accessory attribute modifiers from all accessories in the player's inventory.
 	 *
 	 * @param player the <code>Player</code> to collect from
 	 * @return a <code>Map</code> of attribute modifiers
 	 */
-	public static Map<AttributeModifier, Attribute> collectStandardAttributes(Player player) {
-		Map<AttributeModifier, Attribute> attributeMap = new HashMap<>(5);
+	public static List<AttributeOperation> collectStandardAttributes(Player player) {
+		List<AttributeOperation> attributes = new ArrayList<>(5);
 
 		Optional<ICuriosItemHandler> optional = CuriosApi.getCuriosInventory(player);
 		if (optional.isPresent()) {
 			ICuriosItemHandler itemHandler = optional.get();
-			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItem() instanceof AccessoryItem);
+			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItemHolder().getData(Accessory.ACCESSORY) != null);
 
 			for (SlotResult slotResult : curios) {
-				AccessoryItem accessoryItem = (AccessoryItem) slotResult.stack().getItem();
-				if (!accessoryItem.getStandardAttributeModifiers().isEmpty()) {
-					attributeMap.putAll(accessoryItem.getStandardAttributeModifiers());
+				Accessory accessory = slotResult.stack().getItemHolder().getData(Accessory.ACCESSORY);
+				if (accessory != null) {
+					attributes.addAll(accessory.attributeModifiers());
 				}
 			}
 		}
 
-		return attributeMap;
+		return attributes;
 	}
 
 	/**
 	 * A modified version of {@link AccessoryManager#collectDynamicAttributes(Player)} designed to work with Curios.
-	 * Collect the dynamic accessory attribute modifiers from all {@link AccessoryItem}s in the player's inventory.
+	 * Collect the dynamic accessory attribute modifiers from all accessories in the player's inventory.
 	 *
 	 * @param player the <code>Player</code> to collect from
 	 * @return a <code>Map</code> of attribute modifiers with their target values
 	 */
-	public static Map<Map<AttributeModifier, Attribute>, Double> collectDynamicAttributes(Player player) {
-		Map<Map<AttributeModifier, Attribute>, Double> attributeMap = new HashMap<>(5);
+	public static List<DynamicAttributeOperationInstance> collectDynamicAttributes(Player player) {
+		List<DynamicAttributeOperationInstance> dynamicAttributes = new ArrayList<>(5);
 
 		Optional<ICuriosItemHandler> optional = CuriosApi.getCuriosInventory(player);
 		if (optional.isPresent()) {
 			ICuriosItemHandler itemHandler = optional.get();
-			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItem() instanceof AccessoryItem);
+			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItemHolder().getData(Accessory.ACCESSORY) != null);
 
 			for (SlotResult slotResult : curios) {
-				AccessoryItem accessoryItem = (AccessoryItem) slotResult.stack().getItem();
-				if (!accessoryItem.getDynamicAttributeModifiers().isEmpty()) {
-					attributeMap.putAll(accessoryItem.getDynamicAttributeModifiers());
+				Accessory accessory = slotResult.stack().getItemHolder().getData(Accessory.ACCESSORY);
+				if (accessory != null) {
+					dynamicAttributes.addAll(accessory.dynamicAttributeModifiers());
 				}
 			}
 		}
 
-		return attributeMap;
+		return dynamicAttributes;
 	}
 
 	/**
-	 * A modified version of {@link AccessoryManager#collectMobEffects(Player)} designed to work with Curios.
-	 * Collect the mob effect instances from all {@link AccessoryItem}s in the player's inventory.
+	 * A modified version of {@link AccessoryManager#collectMobEffects(Player)} designed to work with Curios. Collect
+	 * the mob effect instances from all attributes in the player's inventory.
 	 *
 	 * @param player the <code>Player</code> to collect from
 	 * @return a <code>List</code> of mob effect instances
@@ -112,12 +121,12 @@ public class AccessoryBridge {
 		Optional<ICuriosItemHandler> optional = CuriosApi.getCuriosInventory(player);
 		if (optional.isPresent()) {
 			ICuriosItemHandler itemHandler = optional.get();
-			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItem() instanceof AccessoryItem);
+			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItemHolder().getData(Accessory.ACCESSORY) != null);
 
 			for (SlotResult slotResult : curios) {
-				AccessoryItem accessoryItem = (AccessoryItem) slotResult.stack().getItem();
-				if (!accessoryItem.getMobEffects().isEmpty()) {
-					effectList.addAll(accessoryItem.getMobEffects());
+				Accessory accessory = slotResult.stack().getItemHolder().getData(Accessory.ACCESSORY);
+				if (accessory != null) {
+					effectList.addAll(accessory.mobEffectInstances());
 				}
 			}
 		}
@@ -126,19 +135,19 @@ public class AccessoryBridge {
 	}
 
 	/**
-	 * A modified version of {@link AccessoryItem#isAccessoryActive(Player, AccessoryItem)} designed to work with Curios.
-	 * Checks if a given accessory is equipped in a Curios slot.
+	 * A modified version of {@link Accessory#isAccessoryActive(Player, ItemStack)} designed to work with Curios. Checks
+	 * if a given accessory is equipped in a Curios slot.
 	 * <p>
 	 * Because this checks for the first item, it will always be an active one.
 	 */
-	public static boolean isAccessoryActive(Player player, AccessoryItem item) {
+	public static boolean isAccessoryActive(Player player, ItemStack stack) {
 		Optional<ICuriosItemHandler> optional = CuriosApi.getCuriosInventory(player);
 		if (optional.isPresent()) {
 			ICuriosItemHandler itemHandler = optional.get();
-			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItem() == item);
+			List<SlotResult> curios = itemHandler.findCurios(itemStack -> itemStack.getItemHolder().getData(Accessory.ACCESSORY) != null && itemStack.getItem() == stack.getItem());
 
 			// Check if the item has a cooldown
-			if (player.getCooldowns().isOnCooldown(item)) {
+			if (player.getCooldowns().isOnCooldown(stack.getItem())) {
 				return false;
 			}