+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
-This release includes Chinese translations and one bugfix.
+This release targets IW v1.31.0.
 
 ### Feature Changes / Additions
 
-- Added Chinese translations, courtesy of @shenyx110.
-- Add missing Reinforced Depth Charm accessory to the charms tag.
+- Switch Curios implementation back to the official one, now that it is being maintained again
+
+### Bugfixes
+
+- Fix some accessory effects being applied to the player when they shouldn't be